Add files via upload
id text file. Works with the zoltar_project.py file. When we get  the id of the user who mentioned us (our twitter account) we store the id in this txt file and it reads all the tweet posted after it. But this function doesn't work as I have the code 89 error